frontend/requirements.txt
streamlit==1.37.0
pytest==8.3.2
pytest-cov==5.0.0
python-dotenv==1.0.1